feat: implement MQTT client and vending machine service for handling commands

+++ server/domain/vending_machine.go
@@ -8,7 +8,7 @@ type VendingMachineItem struct {
 
 type Transaction struct {
 	ID            string  `json:"id" graphql:"id"`
-	CombinationID int64   `json:"combination_id" graphql:"combination_id"`
+	CombinationID int     `json:"combination_id" graphql:"combination_id"`
 	Price         float64 `json:"price" graphql:"price"`
 	Status        bool    `json:"status" graphql:"status"`
 	TexRef        string  `json:"tex_ref" graphql:"tex_ref"`
